requirements.txt
pandas>=2.0.0
redshift-connector>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
openpyxl>=3.1.0
pyarrow>=12.0.0
